+++ cognite/replicator/events.py
@@ -52,7 +52,7 @@ def create_event(
         asset_ids=replication.get_asset_ids(src_event.asset_ids, src_dst_ids_assets),
         source=src_event.source,
         data_set_id=datasets.replicate(src_client, dst_client, src_event.data_set_id, src_dst_dataset_mapping)
-        if config.get("dataset_support", False)
+        if config and config.get("dataset_support", False)
         else None,
     )
 
@@ -102,7 +102,7 @@ def update_event(
     dst_event.source = src_event.source
     dst_event.data_set_id = (
         datasets.replicate(src_client, dst_client, src_event.data_set_id, src_dst_dataset_mapping)
-        if config.get("dataset_support", False)
+        if config and config.get("dataset_support", False)
         else None,
     )
     return dst_event
@@ -316,15 +316,17 @@ def replicate(
 
     if delete_replicated_if_not_in_src:
         ids_to_delete = replication.find_objects_to_delete_if_not_in_src(events_src, events_dst)
-        client_dst.events.delete(id=ids_to_delete)
-        logging.info(
-            f"Deleted {len(ids_to_delete)} events in destination ({project_dst})"
-            f" because they were no longer in source ({project_src})   "
-        )
+        if ids_to_delete:
+            client_dst.events.delete(id=ids_to_delete)
+            logging.info(
+                f"Deleted {len(ids_to_delete)} events in destination ({project_dst})"
+                f" because they were no longer in source ({project_src})   "
+            )
     if delete_not_replicated_in_dst:
         ids_to_delete = replication.find_objects_to_delete_not_replicated_in_dst(events_dst)
-        client_dst.events.delete(id=ids_to_delete)
-        logging.info(
-            f"Deleted {len(ids_to_delete)} events in destination ({project_dst}) because"
-            f"they were not replicated from source ({project_src})   "
-        )
+        if ids_to_delete:
+            client_dst.events.delete(id=ids_to_delete)
+            logging.info(
+                f"Deleted {len(ids_to_delete)} events in destination ({project_dst}) because"
+                f"they were not replicated from source ({project_src})   "
+            )