+++ cognite/replicator/replication.py
@@ -5,14 +5,14 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 import requests
 from cognite.client import CogniteClient
-from cognite.client.data_classes import Event, FileMetadata, Sequence, TimeSeries
+from cognite.client.data_classes import Event, FileMetadata, Relationship, Sequence, TimeSeries
 from cognite.client.data_classes.assets import Asset
 from cognite.client.data_classes.raw import Row
 
 
 def make_id_object_map(
-    objects: List[Union[Asset, Event, FileMetadata, Sequence, TimeSeries]]
-) -> Dict[int, Union[Asset, Event, FileMetadata, Sequence, TimeSeries]]:
+    objects: List[Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]]
+) -> Dict[int, Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]]:
     """
     Makes a dictionary with the source object id as the key and the object as the value for objects
     that have been replicated.
@@ -31,12 +31,12 @@ def make_id_object_map(
 
 
 def filter_objects(
-    objects: Union[List[Event], List[FileMetadata], List[Sequence], List[TimeSeries]],
+    objects: Union[List[Event], List[FileMetadata], List[Relationship], List[Sequence], List[TimeSeries]],
     src_dst_ids_assets: Dict[int, int],
     skip_unlinkable: bool = False,
     skip_nonasset: bool = False,
-    filter_fn: Optional[Callable[[Union[Event, FileMetadata, Sequence, TimeSeries]], bool]] = None,
-) -> Union[List[Event], List[FileMetadata], List[Sequence], List[TimeSeries]]:
+    filter_fn: Optional[Callable[[Union[Event, FileMetadata, Relationship, Sequence, TimeSeries]], bool]] = None,
+) -> Union[List[Event], List[FileMetadata], List[Relationship], List[Sequence], List[TimeSeries]]:
     """Filters out objects based on their assets and optionally custom filter logic
 
     Args:
@@ -118,7 +118,9 @@ def get_asset_ids(ids: List[int], src_dst_ids_assets: Dict[int, int]) -> Union[L
 
 
 def new_metadata(
-    obj: Union[Asset, Event, FileMetadata, Sequence, TimeSeries], project_src: str, replicated_runtime: int
+    obj: Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries],
+    project_src: str,
+    replicated_runtime: int,
 ) -> Dict[str, Union[int, str]]:
     """
     Copies the objects metadata and adds three new fields to it providing information about the objects replication.
@@ -197,8 +199,8 @@ def restore_fields(
 
 
 def make_objects_batch(
-    src_objects: List[Union[Asset, Event, FileMetadata, Sequence, TimeSeries]],
-    src_id_dst_map: Dict[int, Union[Asset, Event, FileMetadata, Sequence, TimeSeries]],
+    src_objects: List[Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]],
+    src_id_dst_map: Dict[int, Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]],
     src_dst_ids_assets: Dict[int, int],
     create,
     update,
@@ -209,12 +211,12 @@ def make_objects_batch(
     src_dst_dataset_mapping: Dict[int, int],
     config: Dict,
     depth: Optional[int] = None,
-    src_filter: Optional[List[Union[Event, FileMetadata, Sequence, TimeSeries]]] = None,
+    src_filter: Optional[List[Union[Event, FileMetadata, Relationship, Sequence, TimeSeries]]] = None,
     exclude_fields: Optional[List[str]] = None,
 ) -> Tuple[
-    List[Union[Asset, Event, FileMetadata, Sequence, TimeSeries]],
-    List[Union[Asset, Event, FileMetadata, Sequence, TimeSeries]],
-    List[Union[Asset, Event, FileMetadata, Sequence, TimeSeries]],
+    List[Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]],
+    List[Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]],
+    List[Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]],
 ]:
     """
     Create a batch of new objects from a list of source objects or update existing destination objects to their
@@ -270,34 +272,39 @@ def make_objects_batch(
         src_filter_ext_id_set = set()
 
     for src_obj in src_objects:
-        dst_obj = src_id_dst_map.get(src_obj.id)
-        if dst_obj:
-            if not src_obj.last_updated_time or src_obj.last_updated_time > int(dst_obj.metadata["_replicatedTime"]):
-                dst_obj_dump = dst_obj.dump()
-                dst_obj = update(src_obj, dst_obj, src_dst_ids_assets, project_src, replicated_runtime, **kwargs)
+        if hasattr(src_obj, "id"):
+            dst_obj = src_id_dst_map.get(src_obj.id)
+            if dst_obj:
+                if not src_obj.last_updated_time or src_obj.last_updated_time > int(
+                    dst_obj.metadata["_replicatedTime"]
+                ):
+                    dst_obj_dump = dst_obj.dump()
+                    dst_obj = update(src_obj, dst_obj, src_dst_ids_assets, project_src, replicated_runtime, **kwargs)
 
-                if exclude_fields:
-                    dst_obj = restore_fields(dst_obj, dst_obj_dump, exclude_fields)
+                    if exclude_fields:
+                        dst_obj = restore_fields(dst_obj, dst_obj_dump, exclude_fields)
 
-                update_objects.append(dst_obj)
-            else:
-                unchanged_objects.append(dst_obj)
-        elif src_filter:
-            if src_obj.external_id in src_filter_ext_id_set:
-                unchanged_objects.append(src_obj)
+                    update_objects.append(dst_obj)
+                else:
+                    unchanged_objects.append(dst_obj)
+            elif src_filter:
+                if src_obj.external_id in src_filter_ext_id_set:
+                    unchanged_objects.append(src_obj)
+                else:
+                    new_asset = create(src_obj, src_dst_ids_assets, project_src, replicated_runtime, **kwargs)
+                    create_objects.append(new_asset)
             else:
                 new_asset = create(src_obj, src_dst_ids_assets, project_src, replicated_runtime, **kwargs)
                 create_objects.append(new_asset)
         else:
-            new_asset = create(src_obj, src_dst_ids_assets, project_src, replicated_runtime, **kwargs)
+            new_asset = create(src_obj, **kwargs)
             create_objects.append(new_asset)
-
     return create_objects, update_objects, unchanged_objects
 
 
 def retry(
-    function, objects: List[Union[Asset, Event, FileMetadata, Row, Sequence, TimeSeries]], **kwargs
-) -> List[Union[Asset, Event, Sequence, TimeSeries]]:
+    function, objects: List[Union[Asset, Event, FileMetadata, Row, Relationship, Sequence, TimeSeries]], **kwargs
+) -> List[Union[Asset, Event, Relationship, Sequence, TimeSeries]]:
     """
     Attempt to either create/update the objects, if it fails retry creating/updating the objects. This will retry up
     to three times.
@@ -310,7 +317,7 @@ def retry(
         A list of all the objects that were created or updated in CDF.
 
     """
-    ret: List[Union[Asset, Event, FileMetadata, Row, Sequence, TimeSeries]] = []
+    ret: List[Union[Asset, Event, FileMetadata, Row, Relationship, Sequence, TimeSeries]] = []
     if objects:
         tries = 3
         for i in range(tries):
@@ -328,8 +335,8 @@ def thread(
     num_threads: int,
     batch_size: int,
     copy,
-    src_objects: List[Union[Event, FileMetadata, TimeSeries]],
-    src_id_dst_obj: Dict[int, Union[Event, FileMetadata, TimeSeries]],
+    src_objects: List[Union[Event, FileMetadata, Relationship, Sequence, TimeSeries]],
+    src_id_dst_obj: Dict[int, Union[Event, FileMetadata, Relationship, Sequence, TimeSeries]],
     src_dst_ids_assets: Dict[int, int],
     project_src: str,
     replicated_runtime: int,
@@ -458,8 +465,8 @@ def find_objects_to_delete_not_replicated_in_dst(
 
 
 def find_objects_to_delete_if_not_in_src(
-    src_objects: List[Union[Asset, Event, FileMetadata, Sequence, TimeSeries]],
-    dst_objects: List[Union[Asset, Event, FileMetadata, Sequence, TimeSeries]],
+    src_objects: List[Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]],
+    dst_objects: List[Union[Asset, Event, FileMetadata, Relationship, Sequence, TimeSeries]],
 ) -> List[int]:
     """
     Compare the destination and source assets and delete the ones that are no longer in the source.