+++ CHANGELOG.md
@@ -16,6 +16,20 @@ Changes are grouped as follows
 
 - Support for replicating file data
 
+## [1.2.4] - 2023-03-10
+
+## Added
+- Adding replication for relationships
+## Fixed
+- Bug fix on dataset id
+- Fixed replication for sequences
+- Bug fix on deletion of objects not in source or destination
+
+## [1.2.3] - 2023-02-21
+
+## Fixed
+- Bug fix on parameters
+
 ## [1.2.2] - 2023-02-09
 
 ## Fixed